Rename module so that it is installed into the wheel
