+++ python/cudf/cudf/testing/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2020-2025, NVIDIA CORPORATION.
 
-import cudf.testing.narwhals_tests_plugin as narwhals_tests_plugin
+from cudf.testing import narwhals_test_plugin
 from cudf.testing.testing import (
     assert_eq,
     assert_frame_equal,
